requirements.txt
flask
schedule
requests
GitPython
uwsgi